@@ -4,17 +4,11 @@ import { defineConfig } from "vite"
 
 export default defineConfig({
   test: {
-    include: ["./test/**/*.test.{js,mjs,cjs,ts,mts,cts,jsx,tsx}"],
-    exclude: [],
-    globals: true,
-    coverage: {
-      provider: "v8"
-    }
+    include: ["./test/**/*.ts"]
   },
   resolve: {
     alias: {
-      "@effect/docgen/test": path.resolve(__dirname, "/test"),
-      "@effect/docgen": path.resolve(__dirname, "/src")
+      "@effect/docgen": path.resolve(__dirname, "src")
     }
   }
 })
